requirements-init.txt
requests==2.31.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
loguru==0.7.2